Initial working version of the controller
This controller can take a Run with object parameters,
buid, validate and send a CDEvent to a sink.
Example of runs are in the examples folder.
If the parameters are invalid the run will fail and
include why in the reason.
The sink is hardcoded for now, it goes to a knative
broker in cluster URL.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>